requirements.txt
aiofiles==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
azure-ai-translation-text==1.0.0b1
azure-cognitiveservices-speech==1.24.0
azure-core==1.32.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
ecdsa==0.19.0
fastapi==0.95.0
h11==0.14.0
httptools==0.6.4
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
isodate==0.7.2
moviepy==1.0.3
msrest==0.7.1
numpy==2.1.3
oauthlib==3.2.2
pillow==11.0.0
proglog==0.1.10
pyasn1==0.6.1
pydantic==1.10.7
pydub==0.25.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.26.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.21.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
